feat: Adding Windows Testinfra to Packer Build
In order to have Testinfra work with windows packer builds the
configuration needs to support local shell execution. This is
done via the local shell provider from packer. The local shell
provider supports out of the box cmd shell. If you want to use
powershell you have to configure execute_command and corresponding
variables.

It turned out that there is a bug in the provisioner when specifying
a powershell script through the local shell provider. It spawns a
mouse settings dialog when executed and doesn't continue. Turning
to inline scripts worked fine.

https://www.packer.io/docs/provisioners/shell-local.html

Further the configuration needs to support WinRM calls via basic
authentication, which is enabled in the setup.ps1 script.

https://testinfra.readthedocs.io/en/latest/backends.html#winrm

Testinfra's WinRM leverages the pywinrm package

https://github.com/diyan/pywinrm

This commit has been tested successfully on Windows as the packer
host, also the script had to be split as we have a dedicated set
of tests for the Windows image.